% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{hiper_glm}
\usage{
hiper_glm(design, outcome)
}
\arguments{
\item{design}{a design matrix}

\item{outcome}{a vector of outcomes}
}
\value{
An object of class \code{hglm}
}
\description{
hiper_glm
}
